chore: fold RunSlot into the dashboard module
